+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ```
+v0.0.3:
+  date: tbd
+  changes:
+    - Created make_change.c
+    - Updated friendly.c - no longer requires CS50 library installed
+
 v0.0.2:
   date: 2016-01-29
   changes: